Moved get_ruptures in hazardlib
